re-organized some code
