--- pkg/client/client.go
@@ -1,7 +1,0 @@
-package client
-
-type DataflowClient struct {
-	Project  string
-	Location string
-	Prefix   string
-}
