+++ pkg/model/entry.go
@@ -1,4 +1,4 @@
-package models
+package model
 
 import "time"
