+++ pkg/storage/storage.go
@@ -1,4 +1,4 @@
-package interfaces
+package storage
 
 import "time"
